CNAF->Catala input conversion (injection but not surjection)
